% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmfastr.R
\docType{package}
\name{wmfastr}
\alias{wmfastr}
\title{Speedy Dwelltime and User Preference Computations}
\description{
Speedily computes various dwelltime and preference-related
metrics in the context of search sessions.
}
